Dark mode, bugfixes
I fell like there is no patch without writing I fixed certain bugs.
Bluetooth mutliplayer wasn't quite working, because when I told the
camera to jump i made it jump on z axis as well....
Added dark mode! Which is the big change in this patch. Now I have a
button in menu which has a neat animation and then it changes the whole
game to dark mode.
From a programming standpoint I added ChangeUIAccordingToColorMode class
which does what it says. It can change images and texts. If we want them
custom color we need to add the DarkLightColor class to them.
I feel like it is starting to come together.
Also tidied up folders a bit.
TODO: polish bluetooth a little more