GameDones moved to DDOL, delegate color change
GameDones with the grid moved to DontDestroyOnLoad and now you can add
string to DoneGameDisabler on which screens these should be enabled.
Added delegate for colorChange and now i should be faster!
If bluetooth is not working for some reason you should try reverting the
BluetoothDrawGrid.
Also improved preformance by dynamic batching. The signs in game now
have sorting layers.