AI baseworks
Started writin AIScript which I think I'm going to be spending my
upcoming month in.
Wrote a really basic oard evaluation script which calculates point based
on how many signs are next to each other.
I think it will have to be optimised so it only updates the board and
does not calculate it all again every time in a node of alpha beta
heureistic.
Added events in grid for when sign is placed and a sign is removed.
AI is aware of the board.
AI ExpandBoard algorithm probably has to be rewritten if not,
IMAFUCKINGGOD.