AI improvements, defense mechanism
Improved AI.
Now it has a defense mechanism. It checks before the normal evaluations all of the signsInARow and it has a lower percentage chance of skipping it than the normal evaluation. It check for 4s and adds one to an end which is not occupied. It also checks for 3s with no blocks at end and decides which of the two positions is better.

Now while playing the menu buttons animation you can't click buttons which are going out.

Also resolved a bug where when we moved the camera after placing down some signs they moved a tiny bit.