Small improvements: Choose your difficulty panel size increased before gameAI Now signing it to google play is really checked every time the user goes back to menu after 5 secs
Small features:
Implemented sound for: reward screen, borders
Implemeted music. There are currently 20 minutes of music but i hope to add more.

Bugs:
Solved bug where the first clicks on a screen wouldn't be deteced. It was because of the coming back from the settings panel when clicking screen feature, it blocked clicks.
The menu buttons were coming in from bottom left.
Fixed when the back button was pressed while the settings panel was opening the whole screen stayed a bit out of screen.