Features: Added a disabled cell border to the games from now on so they are divided more.
Small features:
Pop up now pops up above settings panel.
Added a little button for the composers.
Now Connecting... panel in Bluetooth is doing a . .. ... . .. ... animation for the points.
Moved settingsbutton in bluetooth because it obstructed the Ready button.

Bugs:
Fixed when we placed  sign ingame then it wouldn't check whether the cell we are trying to put in is disabled or not.
Fixed when we adjusted sound/music in game it would move the camera.
Fixed when the ai places a sign that is out of camera sight it wouldn't move the camera there.