Polish commit #3^43 - 2^10 + 1 Features:
Small features:
AI when out of sight now does not move the camera to be exactly on top of the not seen sign, rather just moves it a bit, so it is visible.
Lowered exp needed for levels above 7. Now the increment is not as steep as it used to be.
Added more icons to achievements in google developer console.

Bugs:
Game in single don't show up only after moving the camera solved.
If settings panel is out in game screen and we click pop-up button the settings panel closes first solved.
In game screen some buttons are above popup wholescreen and in bluetooth connect screen as well solved.
If we switch screens while the settings panel is out the screen doesn't move back fixed.
Pop-up not above buttons in gp and bluetooth connect screens fixed.